Refactor code; Add suggested changes
