+++ circuits-circom/regexes/common/from_regex_v2.circom
@@ -2,7 +2,7 @@ pragma circom 2.1.5;
 
 include "@zk-email/circuits/regexes/regex_helpers.circom";
 
-template FromRegexNew(msg_bytes) {
+template FromRegexV2(msg_bytes) {
 	signal input msg[msg_bytes];
 	signal output out;
