+++ circuits-circom/regexes/paylah/paylah_payer_mobile_num.circom
@@ -2,7 +2,7 @@ pragma circom 2.1.5;
 
 include "@zk-email/circuits/regexes/regex_helpers.circom";
 
-template PaylahPayerIdRegex(msg_bytes) {
+template PaylahPayerMobileNumRegex(msg_bytes) {
 	signal input msg[msg_bytes];
 	signal output out;
