+++ circuits-circom/regexes/paylah/paylah_payment_details.json
