--- circuits-circom/regexes/paylah/paylah_send_amount.json
@@ -1,20 +1,0 @@
-{
-    "parts": [
-        {
-            "is_public": false,
-            "regex_def": "<td>Amount:</td>\r\n"
-        },
-        {
-            "is_public": false,
-            "regex_def": "<td>SGD"
-        },
-        {
-            "is_public": true,
-            "regex_def": "(0|1|2|3|4|5|6|7|8|9|\\.)+"
-        },
-        {
-            "is_public": false,
-            "regex_def": "</td>\r\n"
-        }
-    ]
-}
