--- circuits-circom/test/mocks/paylah/test_paylah_all.circom
@@ -1,3 +1,0 @@
-include "../../../regexes/paylah/paylah_all_regex.circom";
-
-component main = PaylahAllRegex(189);
