+++ circuits-circom/test/mocks/paylah/test_paylah_mobile_num.circom
@@ -1,0 +1,3 @@
+include "../../../regexes/paylah/paylah_payer_mobile_num.circom";
+
+component main = PaylahPayerMobileNumRegex(62);
