--- circuits-circom/test/mocks/paylah/test_paylah_payee_id.circom
@@ -1,3 +1,0 @@
-include "../../../regexes/paylah/paylah_payee_id.circom";
-
-component main = PaylahPayeeIdRegex(68);
