--- circuits-circom/test/mocks/paylah/test_paylah_payer_id.circom
@@ -1,3 +1,0 @@
-include "../../../regexes/paylah/paylah_payer_id.circom";
-
-component main = PaylahPayerIdRegex(62);
