+++ circuits-circom/test/mocks/paylah/test_paylah_payment_details.circom
@@ -1,0 +1,3 @@
+include "../../../regexes/paylah/paylah_payment_details.circom";
+
+component main = PaylahPaymentDetailsRegex(189);
