--- circuits-circom/test/mocks/paylah/test_paylah_send_amount.circom
@@ -1,3 +1,0 @@
-include "../../../regexes/paylah/paylah_send_amount.circom";
-
-component main = PaylahSendAmountRegex(47);
